requirements.txt
fastapi==0.115.12
Flask==3.1.3
PyInstaller==6.13.0
PyPDF2==3.0.1
pywebview==5.4
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.34.2
